rename consumption emssions file and modify test
