+++ data/issues/consumption/consumption_emissions.py
